djangoapp/requirements.txt
asgiref==3.7.2
Django>=5.0,<5.1
djangorestframework==3.14
djangorestframework-gis>=1.0
django-cors-headers==4.1.0
sqlparse==0.4.4
tzdata==2024.1
psycopg2-binary==2.9.9